Create file reader.go
